beautifulsoup-headless/requirements.txt
Flask>=2.0.0
psycopg2>=2.9.0
python-dotenv>=0.19.0
requests>=2.26.0
beautifulsoup4>=4.10.0